Python Sockets Communication
In this project I used a client wich connects to two servers. It will
provide two numbers for the first server,it closes the first server
connection and oppens the connection with the second server, it provides
another two numbers for the second server. After receiveing the numbers,
each server will become a client and send the sum and the multiplication
of the numbers they received to a third server. This last server will
then print out the result. This project helped me understand more about
socket programming.